Improve Vision Viewer and Network Tables Output
Do not add Blur filter, this is done by the camera lens, much more
processing efficient.

Combine FoundTarget (boolean) and FrameCount so that FrameCount only
increments when a new image with targets is found rather than every time
an image is found. This way the RoboRio will only use the imformation from
Distance/Angle when the image is both new and good.

Begin checking for existance of NetworkTable which implies RoboRio
connectivity, which allows us to both save time (computing Distance etc is
useless if RoboRio isn't there to use the info) and prevent errors when using
the static VisionTool while disconnected (on saved images).

Begin cloning inputs to methods that take a Mat as an argument, so that
other methods may use the original image in it's unmodified form.