Add Explanation Image for OffCenter Angle Calc
Add image to /res/

Update .gitignore to ignore /wiki/ so that the wiki pages repository may
be cloned within the 2016-Vision repo/directory and not track the files in
there, which would be duplicating these files across repos.

By placing the wiki repo within the main vision repo, links to files
(which use relative dirs) within the main repo (such as to /res/) are
preserved both on github and in local directory markdown rendering.